reading edited pdf file with image

+++ components/Sign/functions.js
@@ -1,16 +1,22 @@
+import { encode } from "base-64";
 import * as FileSystem from "expo-file-system";
+import { Alert } from "react-native";
 import ReactNativeBlobUtil from "react-native-blob-util";
+import RNFS from "react-native-fs";
 
 export const selectSignature = async (
   signatureFilePath,
-  navigation,
   setInputSignature,
-  setSelectedSignaturePath
+  setSelectedSignaturePath,
+  setSignatureBase64Data
 ) => {
   // const info = await FileSystem.getInfoAsync(signatureFilePath);
   // ReactNativeBlobUtil.ios.openDocument(signatureFilePath);
   // alert(info.uri);
   // navigation.navigate("DragSignature");
+
+  const signatureBase64 = await RNFS.readFile(signatureFilePath, "base64");
+  setSignatureBase64Data(signatureBase64);
 
   setSelectedSignaturePath(() => signatureFilePath);
   setInputSignature((curr) => !curr);
@@ -21,45 +27,44 @@ export const deleteSignature = (
   signatureList,
   setSignatureList
 ) => {
-  const path = fileWtPath.split("//");
-  console.log("path:", path);
+  return Alert.alert(
+    "Are your sure?",
+    "Are you sure you want to delete this signature?",
+    [
+      {
+        text: "Yes",
+        onPress: () => {
+          const path = fileWtPath.split("//");
+          console.log("path:", path);
 
-  ReactNativeBlobUtil.fs
-    .unlink(path[1])
-    .then(() => {
-      console.log("Deleted File from - ", path[1]);
-    })
-    .catch((err) => console.log(err));
+          ReactNativeBlobUtil.fs
+            .unlink(path[1])
+            .then(() => {
+              console.log("Deleted File from - ", path[1]);
+            })
+            .catch((err) => console.log(err));
 
-  console.log("signatureList", signatureList);
+          console.log("signatureList", signatureList);
 
-  // Updating signature array list for the UI
-  const updatedSignatureList = signatureList.filter((signature) => {
-    console.log("signature", signature);
-    return signature !== fileWtPath;
-  });
+          // Updating signature array list for the UI
 
-  console.log("updatedSignatureList", updatedSignatureList);
+          const updatedSignatureList = signatureList.filter((signature) => {
+            console.log("signature", signature);
+            return signature !== fileWtPath;
+          });
 
-  setSignatureList(updatedSignatureList);
+          console.log("updatedSignatureList", updatedSignatureList);
 
-  // return Alert.alert(
-  //   "Are your sure?",
-  //   "Are you sure you want to delete this signature?",
-  //   [
-  //     // The "Yes" button
-  //     {
-  //       text: "Yes",
-  //       onPress: () => {
+          setSignatureList(updatedSignatureList);
+        },
+      },
 
-  //       },
-  //     },
-
-  //     {
-  //       text: "No",
-  //     },
-  //   ]
-  // );
+      {
+        text: "No",
+        onPress: () => {},
+      },
+    ]
+  );
 };
 
 export const hideSignatures = (setSignatureList) => {
@@ -81,3 +86,17 @@ export async function displayStoredSignatures(setSignatureList) {
 
   setSignatureList([...updatedSignatureList]);
 }
+
+export const uint8ToBase64Conversion = (u8Arr) => {
+  const CHUNK_SIZE = 0x8000; // arbitrary number
+  let index = 0;
+  const length = u8Arr.length;
+  let result = "";
+  let slice;
+  while (index < length) {
+    slice = u8Arr.subarray(index, Math.min(index + CHUNK_SIZE, length));
+    result += String.fromCharCode.apply(null, slice);
+    index += CHUNK_SIZE;
+  }
+  return encode(result);
+};
